MetricsMonitoringAndCollection/services/data-consumer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
influxdb-client==1.38.0
pydantic==2.5.0